+++ CHANGELOG.md
@@ -2,8 +2,13 @@
 # Change Log
 
 ## [In Development](https://github.com/ppfeufer/wp-esi-client/tree/development)
-[Full Changelog](https://github.com/ppfeufer/wp-esi-client/compare/v20181202...development)
+[Full Changelog](https://github.com/ppfeufer/wp-esi-client/compare/v20181226...development)
 - in development
+
+## [v20181226](https://github.com/ppfeufer/wp-esi-client/tag/v20181226) - 2018-12-26
+[Full Changelog](https://github.com/ppfeufer/wp-esi-client/compare/v20181202...v20181226)
+### Added
+- Model for universe/categories/category_id added » https://esi.evetech.net/ui#/Universe/get_universe_categories_category_id
 
 ## [v20181202](https://github.com/ppfeufer/wp-esi-client/tag/v20181202) - 2018-12-02
 [Full Changelog](https://github.com/ppfeufer/wp-esi-client/compare/v20181125...v20181202)